Changing the number of features means manually changing the input_size param in the model config.
So I changed the order of code in trainer.py to initialise the DataLoaders earlier, and run one load to get the feature dimensions.

Then I created set_config_param in utils.py to add the input_size param to the model config.

Now input_size is not needed in the model configs. If it is present, it will be overwritten anyay.