Fixed the training code to handle the non cross_validation case:
Previously, the output type in get_dataset() in trainer.py was different depending if cross_validation was True or False. True: Dict of LandmarkDatasets, False: Dict of DataLoaders. Then, train_epoch() in train.py expected *datasets* as a Dict of DataLoaders, but the type hint expected a Dict of LandmarkDatasets. So I fixed get_dataset() to give the same type of output either way.

In train.py, I fixed the train_epoch() code to handle the Dict of LandmarkDatasets in a similar way to train_epoch_fold(). Also added type hints and unified the code style for the 2 functions.