Added more options to angles_estimator, and set the angle_type to *clockwise_func* in the config.
Previously, the estimator only used the dot product and arccos to get the angle. This gets the smallest angle between two lines. So 30 or 330 degrees clockwise both become -> 30.

This is problematic for 2 reasons:
1. We want to know the orientation of the two lines we are getting the angle of.
2.  This makes the range of angles [0, π] radians or [0, 180] degrees. For our features, we should use the *func* config, which gets the sin() & cos() of the angle to give us numbers between [-1, 1] that are more appropriate for an ML model. However if the angle is restricted to [0, π], then the range of sin(angle) is [0, 1].

One option is to scale this between [-1,1]. but considering the other issue too, it is probably more sensible to just use the full clockwise angle.